requirement.txt
mysql-connector-python>=9.0.0
python-dotenv>=1.0.1
tzdata>=2024.1
